# perf: revamp `symbolic.EvalBoard`