--- a/prover/circuits/internal/utils.go
+++ b/prover/circuits/internal/utils.go
@@ -40,7 +40,7 @@ func AssertIsLessIf(api frontend.API, cond, a, b frontend.Variable) {
 	api.AssertIsLessOrEqual(a_, b_)
 }
 
-func SliceToTable(api frontend.API, slice []frontend.Variable) *logderivlookup.Table {
+func SliceToTable(api frontend.API, slice []frontend.Variable) logderivlookup.Table {
 	table := logderivlookup.New(api)
 	for i := range slice {
 		table.Insert(slice[i])
@@ -271,7 +271,7 @@ func (s VarSlice) Range(api frontend.API) *Range {
 func Concat(api frontend.API, maxLinearizedLength int, slices ...VarSlice) Slice[frontend.Variable] {
 
 	res := Slice[frontend.Variable]{make([]frontend.Variable, maxLinearizedLength), 0}
-	var outT *logderivlookup.Table
+	var outT logderivlookup.Table
 	{ // hint
 		inLen := 2 * len(slices)
 		for i := range slices {
@@ -414,7 +414,7 @@ func ChecksumSubSlices(api frontend.API, hsh snarkHash.FieldHasher, slice []fron
 		lastElems[i] =
 			plonk.EvaluateExpression(api, e, endpointsR.InRange[i], 0, -1-len(slice), 1, len(slice))
 	}
-	var lastElemsT *logderivlookup.Table
+	var lastElemsT logderivlookup.Table
 	if len(slice) > 1 {
 		lastElemsT = SliceToTable(api, lastElems)
 		lastElemsT.Insert(len(slice)) // in case subEndPoints is tight
@@ -730,7 +730,7 @@ func PartitionSlice(api frontend.API, s []frontend.Variable, selectors []fronten
 		panic(err)
 	}
 
-	subsT := make([]*logderivlookup.Table, len(subs))
+	subsT := make([]logderivlookup.Table, len(subs))
 	for i := range subs {
 		copy(subs[i], subsGlued[:len(subs[i])])
 		subsGlued = subsGlued[len(subs[i]):]