--- a/prover/circuits/prove.go
+++ b/prover/circuits/prove.go
@@ -86,7 +86,6 @@ func ProveCheck(setup *Setup, assignment frontend.Circuit, opts ...any) (plonk.P
 			assignment,
 			assignment,
 			setup.Circuit.Field(),
-			test.WithBackendProverOptions(proverOpts...),
 		)
 		return nil, fmt.Errorf("while running the plonk prover: %w", errDetail)
 	}