--- a/prover/maths/common/smartvectors/arithmetic_basic_ext.go
+++ b/prover/maths/common/smartvectors/arithmetic_basic_ext.go
@@ -110,6 +110,7 @@ func LinearCombinationExt(vecs []SmartVector, x fext.Element, p ...mempool.MemPo
 			acc[i].Add(&acc[i], &tmpF)
 		}
 	}
+
 	// Computes the polynomial operation separately on the const,
 	// windows and regular and the aggregate the results at the end.
 	// The computation is done following horner's method.