--- a/prover/maths/field/element.go
+++ b/prover/maths/field/element.go
@@ -11,7 +11,7 @@ import (
 )
 
 type Element = koalabear.Element
-type Vector []Element
+type Vector = koalabear.Vector
 type Octuplet [8]Element
 
 const (