… .github/workflows/prover-testing.yml
… prover/crypto/poseidon2/poseidon2_test.go
… prover/crypto/ringsis/ringsis_test.go
… prover/go.sum
… prover/protocol/compiler/standard_benchmark_test.go
… prover/symbolic/evaluation_board.go
… prover/symbolic/evaluation_ext.go
… prover/symbolic/evaluation_test.go
… prover/symbolic/expression.go
… prover/symbolic/expression_test.go
… prover/symbolic/polyeval.go