Most of this PR is in `evaluation.go` . 
It gets rid of the multiple `Evaluate` `EvaluateMixed` and `EvaluatedExt` similar functions to keep only one.

The `boardAssignment` is replaced by the `evaluation` struct that allocate the needed memory for each go routine before looping on the chunks it needs to solve. The new evaluation logic also bypasses the `arithmetic_ext / _ mixed etc`.go files. The inputs are smart vectors, they are written in the pre-allocated arrays in the `evaluation` struct, which account for their type (constant, rotated etc). 

Since memory is known in advance for each chunk (i.e. set of rows we process in a go routine), there is no need for the memory pool either.

Perf against `dev-small-fields` for realistic segment:

```
benchmark                                  old ns/op        new ns/op        delta
BenchmarkCompiler/realistic-segment-96     184955073455     156563834619     -15.35%

benchmark                                  old allocs      new allocs      delta
BenchmarkCompiler/realistic-segment-96     22627949643     22340680827     -1.27%

benchmark                                  old bytes         new bytes        delta
BenchmarkCompiler/realistic-segment-96     1071664302168     800542197232     -25.30%
```

(this is without @YaoJGalteland other branches on poseidon2 bytes consumtions, fft domains, avx512 method usage etc).